Makefile: Allow `PLATFORM_VERSION` variable to be changed in `make` call
In commit aa990fe ("Makefile: Set `PLATFORM_VERSION` to 10"), a new
variable called `PLATFORM_VERSION` is defined directly in Makefile.
However, the variable was set with "simply expanded" sign (":="), which
will set the value even if it has been set from another `make` call.
Instead, we can use "?=" for the assignment sign. With this, Linux build
system will not override the said variable.

Change the assignment sign to "?=" to allow `PLATFORM_VERSION` variable
to be changed.

Change-Id: Ibd08e7a1bc5d0c27b1424a2d74b74d03b1f306d1
Signed-off-by: Hendra Manudinata <saya@hendra-manudinata.my.id>